Combine Share and Manage Access UIs
Combine the UI elements for sharing and managing access to a book into a single UI component.
